A classe da CurrencyApi agora está em um único arquivo
-- Os métodos da 'self' da currencyApi foram movidos para o arquivo de definição da classe da currencyApi, então todos os métodos dela agora estão em um único arquivo
-- Isso foi feito pq ela não está grande o suficiente para o nível de confusão que métodos em outros arquivos causa justificar o benefício de uma classe com vários arquivos menores
