--- src/currencyApi/self/index.ts
@@ -1,2 +1,0 @@
-export * from './create_accounts'
-export * from './withdraw'
